refactor(api-testing-core): share non-JSON failure body logging
- Extract Shared stderr preview helper into api-testing-core cli_io.
- Reuse Helper in api-rest and api-grpc call failure paths.
- Add Helper coverage tests and keep grpc call characterization checks.
